Add skeleton for new modularized client and copy implementations over
