--- src/home_robot_hw/home_robot_hw/client/__init__.py
@@ -1,1 +1,0 @@
-from .stretch_hw_interface import StretchClientApi
