--- src/home_robot_hw/home_robot_hw/client/camera.py
@@ -1,3 +1,0 @@
-
-class StretchCameraClient:
-    pass
