+++ src/home_robot_hw/home_robot_hw/stretch_client/__init__.py
@@ -1,0 +1,1 @@
+from .api import StretchClient
